fix: change unsignedBigInteger to unsignedInteger for product_image_id
Coze-Commit-Type: user
Coze-User-ID: 4337991133177372
Coze-Conversation-ID: 4912498

+++ packages/Webkul/Product/src/Database/Migrations/2024_01_01_000001_create_product_customization_tables.php
@@ -14,7 +14,7 @@ return new class extends Migration
         // 商品图片可打印区域表
         Schema::create('product_image_print_areas', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('product_image_id');
+            $table->unsignedInteger('product_image_id');
             $table->foreign('product_image_id')->references('id')->on('product_images')->onDelete('cascade');
             $table->string('name')->nullable()->comment('区域名称');
             $table->decimal('x', 8, 4)->default(0)->comment('X坐标百分比');
